+++ CS5200Project/src/main/java/game/model/Equipments.java
@@ -9,7 +9,7 @@ public class Equipments extends Items {
 			String itemName,
 			int level,
 			int maxStackSize,
-			float price,
+			double price,
 			int requiredLevel
 			) {
 		super(itemID, itemName, level, maxStackSize, price);
